Thread::Queue: Disallow size 1
In d0b3afa83dbdf2fe711a865a64dcbab5d8325bc6, for statistical inference commands the size of the shuffling matrix queue is set based on the number of processing threads, to reduce memory usage. This however fails to execute in the case -nthreads 1. Rather than identifying the blocking condition with the Queue class, this change simply ensures that the size of the queue is always at least 2 items.